+++ packages/fuel-gauge/src/transaction.test.ts
@@ -77,9 +77,9 @@ describe('Transaction', () => {
 
     await request.estimateAndFund(fundedWallet);
 
-    const tx = await fundedWallet.sendTransaction(request);
+    const { waitForResult } = await fundedWallet.sendTransaction(request);
 
-    const { isStatusSuccess } = await tx.waitForResult();
+    const { isStatusSuccess } = await waitForResult();
 
     // Wait for message status to update
     await sleep(1000);
@@ -128,9 +128,9 @@ describe('Transaction', () => {
 
     await request.estimateAndFund(fundedWallet);
 
-    const tx = await fundedWallet.sendTransaction(request);
+    const { waitForResult } = await fundedWallet.sendTransaction(request);
 
-    const { isStatusSuccess } = await tx.waitForResult();
+    const { isStatusSuccess } = await waitForResult();
 
     // Wait for message status to update
     await sleep(1000);
@@ -167,10 +167,10 @@ describe('Transaction', () => {
 
     await request.estimateAndFund(sender);
 
-    const tx = await sender.sendTransaction(request, {
+    const { waitForResult } = await sender.sendTransaction(request, {
       enableAssetBurn: true,
     });
-    const { isStatusSuccess } = await tx.waitForResult();
+    const { isStatusSuccess } = await waitForResult();
     expect(isStatusSuccess).toEqual(true);
     expect(request.outputs).to.not.contain(
       expect.objectContaining({
@@ -255,10 +255,10 @@ describe('Transaction', () => {
 
     await request.estimateAndFund(owner);
 
-    const tx = await owner.sendTransaction(request, {
+    const { waitForResult } = await owner.sendTransaction(request, {
       enableAssetBurn: true,
     });
-    const { isStatusSuccess } = await tx.waitForResult();
+    const { isStatusSuccess } = await waitForResult();
     expect(isStatusSuccess).toEqual(true);
     expect(request.outputs).to.not.contain(
       expect.objectContaining({