+++ packages/versions/src/lib/getBuiltinVersions.ts
@@ -2,8 +2,8 @@ import type { Versions } from './types';
 
 export function getBuiltinVersions(): Versions {
   return {
-    FORC: '0.66.6',
+    FORC: '0.66.7',
     FUEL_CORE: '0.40.4',
-    FUELS: '0.98.0',
+    FUELS: '0.99.0',
   };
 }