+++ packages/crypto/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 0.99.0
+
+### Patch Changes
+
+- Updated dependencies [d1825c9]
+- Updated dependencies [b3bb765]
+  - @fuel-ts/utils@0.99.0
+  - @fuel-ts/errors@0.99.0
+
 ## 0.98.0
 
 ### Patch Changes